Removed acceleration
I originally planned to have this module handle bullets with weird
physics, such as ones with a separate x and y velocity instead of a
speed and angle. I decided to keep this module as lean as possible
instead. Everything can be done with position, speed, and direction. You
just need another node, or a gdscript script, to maintain a list of
bullets with unique properties, and modify the speed/position/angle in
_physics_process.